EcoScanner/requirements.txt
fastapi
uvicorn[standard]
requests
Pillow
pyzbar
python-multipart